Create Are Silver Travertine Tiles Better Than Other Alternatives? Know How!
Travertine is a kind of Limestone formed by naturally occurring mineral deposits in springs. Calcite is a geological stone consisting of calcium carbonate. Travertine tiles have special qualities, which are mostly due to the addition of minerals to calcite. It is one of the most versatile components due to its variety of finishes.